requirements_streamlit.txt
streamlit==1.28.1
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
